Added new test file for #25
Added a new file ExecutionTest.java which now holds the test cases of
the class Execution instead of the ServerTest class
